requirements.txt
# AI-Generated Voice Detection API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
google-genai>=1.0.0
